Deacidify adminlog
